// src/app/main-modules/form-builder2/model/enum/element-type.ts
export enum ElementType {
  TEXT_FIELD = 'TEXT_FIELD' as any,
  CHECK_BOX = 'CHECK_BOX' as any,
  TEXT_AREA = 'TEXT_AREA' as any,
  COMBO_BOX = 'COMBO_BOX' as any,
  RADIO_BUTTON = 'RADIO_BUTTON' as any,
  NUMERICAL = 'NUMERICAL' as any,
  DATE = 'DATE' as any,
  TIME = 'TIME' as any,
  MATRIX = 'MATRIX' as any,
  STAR_RATING = 'STAR_RATING' as any,
  FILE_ATTACH = 'FILE_ATTACH' as any,
}